Project rename
* Rename python package from `ansible_events_ui` to `eda-server`.
* Fix all absolute imports.
* Fix alembic config.
* Fix project metadata:
    - Update `author` and `author_email`.
* Replace `AE_` environment variables prefix with `EDA_`.
