+++ alembic.ini
@@ -2,7 +2,7 @@
 
 [alembic]
 # path to migration scripts
-script_location = ansible_events_ui/db/migrations
+script_location = eda_server/db/migrations
 
 # template used to generate migration file names; The default value is %%(rev)s_%%(slug)s
 # Uncomment the line below if you want the files to be prepended with date and time
@@ -36,10 +36,10 @@ timezone = UTC
 # sourceless = false
 
 # version location specification; This defaults
-# to ansible_events_ui/db/migrations/versions.  When using multiple version
+# to eda_server/db/migrations/versions.  When using multiple version
 # directories, initial revisions must be specified with --version-path.
 # The path separator used here should be the separator specified by "version_path_separator" below.
-# version_locations = %(here)s/bar:%(here)s/bat:ansible_events_ui/db/migrations/versions
+# version_locations = %(here)s/bar:%(here)s/bat:eda_server/db/migrations/versions
 
 # version path separator; As mentioned above, this is the character used to split
 # version_locations. The default within new alembic.ini files is "os", which uses os.pathsep.
