+++ eda_server/main.py
@@ -6,7 +6,7 @@ from .config import load_settings
 def main():
     settings = load_settings()
     uvicorn.run(
-        "ansible_events_ui.app:create_app",
+        "eda_server.app:create_app",
         factory=True,
         host=settings.host,
         port=settings.port,
