+++ setup.cfg
@@ -1,9 +1,9 @@
 [metadata]
-name = ansible_events_ui
+name = eda-server
 version = 0.1.0
-author = Ben Thomasson
-author_email = ben.thomasson@gmail.com
-url = https://github.com/benthomasson/ansible-events-ui
+author = Red Hat, Inc.
+author_email = info@ansible.com
+url = https://github.com/ansible/eda-server
 license = Apache-2.0
 
 [options]
@@ -28,7 +28,7 @@ install_requires =
 [options.entry_points]
 # TODO(cutwater): Maybe rename it as ansible-events-api \ ansible-events-server?
 console_scripts =
-    ansible-events-ui = ansible_events_ui.main:main
+    eda-server = eda_server.main:main
 
 
 [flake8]
