+++ tests/integration/api/test_inventory.py
@@ -4,7 +4,7 @@ from fastapi import status as status_codes
 from httpx import AsyncClient
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from ansible_events_ui.db import models
+from eda_server.db import models
 
 TEST_RULESET_SIMPLE = """
 ---
